refactor: separa classes Carta e Cadastro em arquivos diferentes
